refactor(tools): split skills_tools.py into subprocess helpers, CLI runners, and market gold tools
- Create src/tools/_subprocess.py for shared CLI/shell execution helpers and whale tracker parsing
- Create src/tools/runners.py to isolate the 11 command-runner @tool wrappers (e.g. comex, macro, news)
- Create src/tools/market/gold.py for explain_price_anomalies and run_risk_governor_analysis
- Clean up src/tools/skills_tools.py by removing runner/gold function bodies and duplicate tool lists
- Re-export all refactored tools from skills_tools.py to preserve import compatibility across all agents and tests
