+++ src/tools/skills_tools.py
@@ -1,226 +1,53 @@
 """
 src/tools/skills_tools.py
 ─────────────────────────
-LangChain tool wrappers around the core skills and scripts of the Mosaic-agent.
-Allows the ReAct agent to run goldbees reports, macro scanning, signal aggregator,
-ETF news sentiment, DSP Multi-Asset imports, and risk governor calculations.
+Core skill tools for the Mosaic-agent ReAct agents.
+
+What lives here: general-purpose tools that don't fit a tighter domain —
+ClickHouse query, symbol import, iNAV lookup, premium alerts, deep-dive,
+and the shared subprocess helpers (_run_cmd, _run_cmd_streaming).
+
+Shell-command runners (run_goldbees_pipeline, run_macro_scanner, …) →
+    src/tools/runners.py
+
+Gold/GARCH domain tools (explain_price_anomalies, run_risk_governor_analysis) →
+    src/tools/market/gold.py
+
+Both modules are re-exported here for backward compatibility.
 """
 
 from __future__ import annotations
 
 import os
-import re
 import sys
-import subprocess
-from langchain_core.tools import tool
 import logging
+from langchain_core.tools import tool
+
+from src.tools._subprocess import (  # shared helpers — no circular dep
+    PROJECT_ROOT,
+    _run_cmd,
+    _run_cmd_streaming,
+    _clean_terminal_output,
+)
 
 logger = logging.getLogger(__name__)
 
-PROJECT_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", ".."))
-
-
-def _clean_terminal_output(text: str) -> str:
-    """Removes terminal box-drawing characters and excessive blank lines to make tool output easier for LLMs to read."""
-    # Characters to remove or replace
-    replacements = {
-        "█": "#", "░": ".", "▒": ".", "▓": "#",
-        "■": "*", "▲": "^", "▼": "v"
-    }
-    
-    lines = []
-    for line in text.splitlines():
-        # Specific replacements
-        for char, replacement in replacements.items():
-            line = line.replace(char, replacement)
-        
-        # Clean all Unicode box-drawing characters (U+2500 to U+257F block)
-        cleaned_chars = []
-        for char in line:
-            val = ord(char)
-            if 0x2500 <= val <= 0x257F:
-                if char in ("─", "━", "═", "┄", "┅", "┈", "┉", "╌", "╍"):
-                    cleaned_chars.append("-")
-                else:
-                    cleaned_chars.append("")
-            else:
-                cleaned_chars.append(char)
-        line = "".join(cleaned_chars)
-        
-        # Strip trailing/leading spaces
-        line = line.strip()
-        
-        # Skip empty lines if they were just box drawings or blank
-        if line and not all(c in "-_ " for c in line):
-            lines.append(line)
-            
-    return "\n".join(lines)
-
-
-def _run_cmd(args: list[str]) -> str:
-    """Helper to run a command via subprocess from the project root with the correct Python interpreter."""
-    env = os.environ.copy()
-    env["ALLOW_LOCAL_RUN"] = "1"
-
-    # Ensure project root is in PYTHONPATH
-    if "PYTHONPATH" in env:
-        env["PYTHONPATH"] = PROJECT_ROOT + os.pathsep + env["PYTHONPATH"]
-    else:
-        env["PYTHONPATH"] = PROJECT_ROOT
-
-    cmd = [sys.executable] + args
-    try:
-        res = subprocess.run(
-            cmd,
-            cwd=PROJECT_ROOT,
-            capture_output=True,
-            text=True,
-            env=env,
-            timeout=300,
-        )
-        output = res.stdout
-        if res.stderr:
-            output += "\n--- STDERR ---\n" + res.stderr
-        return _clean_terminal_output(output)
-    except Exception as e:
-        return f"Error executing command {' '.join(cmd)}: {e}"
-
-
-# Pre-compiled ANSI escape code stripper used by streaming output
-import re as _re
-_ANSI_STRIP_RE = _re.compile(r"\x1b\[[0-9;]*[a-zA-Z]|\x1b[()][A-Z0-9=><]|\x1b[ABCDEF78]")
-
-
-def _run_cmd_streaming(args: list[str]) -> str:
-    """
-    Like _run_cmd but prints each output line to the terminal as the subprocess
-    runs, giving live progress. Used for long-running operations like data import.
-    Returns the full collected output as a string when done.
-    """
-    env = os.environ.copy()
-    env["ALLOW_LOCAL_RUN"] = "1"
-    env["NO_COLOR"] = "1"       # tell Rich / Typer inside subprocess to skip ANSI codes
-    env["TERM"] = "dumb"        # further signal: no colour support
-    if "PYTHONPATH" in env:
-        env["PYTHONPATH"] = PROJECT_ROOT + os.pathsep + env["PYTHONPATH"]
-    else:
-        env["PYTHONPATH"] = PROJECT_ROOT
-
-    cmd = [sys.executable] + args
-    collected: list[str] = []
-
-    try:
-        proc = subprocess.Popen(
-            cmd,
-            cwd=PROJECT_ROOT,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.STDOUT,
-            text=True,
-            env=env,
-            bufsize=1,          # line-buffered
-        )
-        for raw_line in iter(proc.stdout.readline, ""):
-            # Strip ANSI codes left over despite NO_COLOR
-            stripped = _ANSI_STRIP_RE.sub("", raw_line).rstrip()
-            # Apply box-drawing cleanup (shared with _run_cmd)
-            cleaned = _clean_terminal_output(stripped)
-            if cleaned:
-                sys.stdout.write(f"  {cleaned}\n")
-                sys.stdout.flush()
-                collected.append(cleaned)
-        proc.wait()
-        rc = proc.returncode
-    except Exception as exc:
-        return f"Import error: {exc}\n" + "\n".join(collected)
-
-    result = "\n".join(collected) if collected else "Import completed (no output)."
-    if rc != 0:
-        result += f"\n[Process exited with code {rc}]"
-    return result
-
-
-@tool
-def run_goldbees_pipeline() -> str:
-    """
-    Run the GOLDBEES pipeline report script.
-    This prints the pre-baked gold recommendation block, including ML prediction probability,
-    Risk Governor and Kelly weights, and iNAV premium.
-    Use this when asked for GOLDBEES recommendation, today's gold signal, or what to do with GOLDBEES.
-    """
-    return _run_cmd(["src/scripts/goldbees_report.py"])
-
-
-@tool
-def run_daily_signal_composite(save: bool = True) -> str:
-    """
-    Run the composite signal aggregator to compute scores for all 18 ETFs.
-    Combines macro events, news sentiment, NAV Z-scores, FII/DII flows, ML predictions, and anomaly regimes.
-    Use this when asked for ETF buy/sell recommendations, composite scores, or signal dashboard.
-    """
-    args = ["src/main.py", "signals"]
-    if save:
-        args.append("--save")
-    return _run_cmd(args)
-
-
-@tool
-def run_macro_scanner(max_themes: int = 3) -> str:
-    """
-    Scan live macro and geopolitical events, mapping their directional impact to Indian ETFs.
-    Use this when asked about active macro themes, geopolitical risks, crude oil/INR shocks, or Fed/RBI policy events.
-    """
-    return _run_cmd(["src/main.py", "macro", "--max", str(max_themes)])
-
-
-@tool
-def run_etf_news_sentiment(max_articles: int = 3, save: bool = True) -> str:
-    """
-    Fetch and tag the latest news articles by ETF category with sentiment scores.
-    Use this when asked for ETF news, sentiment trends, or saving ETF news.
-    """
-    args = ["src/main.py", "etf-news", "--max", str(max_articles)]
-    if save:
-        args.append("--save")
-    return _run_cmd(args)
-
-
-@tool
-def run_dsp_multi_asset_importer() -> str:
-    """
-    Import and backfill DSP Multi Asset Allocation Fund holdings history into the ClickHouse database.
-    Use this when asked to import or update DSP holdings data.
-    """
-    return _run_cmd(["src/scripts/dsp/import_all_dsp_equity.py"])
-
-
-@tool
-def run_data_engineering_importer(category: str = "etfs,stocks,mf,fii_dii,cot,fx_rates,inav", full: bool = False, symbol: str = "") -> str:
-    """
-    Trigger the historical ClickHouse data engineering pipeline to import and sync BULK data.
-    Use ONLY for bulk category imports — e.g. "import all stocks", "refresh ETFs", "sync everything".
-
-    IMPORTANT: If the user names a SPECIFIC symbol (e.g. "import ADVENZYMES", "refresh GOLDBEES"),
-    do NOT use this tool — use `import_symbol_data(symbol)` instead.
-    This tool imports ALL symbols in a category, which is slow and wasteful when only one is needed.
-
-    If called with a symbol anyway, it will auto-redirect to import_symbol_data for that symbol.
-
-    Args:
-        category: Comma-separated list of categories to import.
-                  Valid values: etfs, stocks, mf, fii_dii, cot, fx_rates, inav.
-        full: If True, performs a full backfill ignoring watermarks.
-        symbol: (optional) If a specific symbol is provided, redirects to import_symbol_data.
-    """
-    # Safety net: if caller passed a specific symbol, redirect to per-symbol import
-    if symbol and symbol.strip():
-        return import_symbol_data_impl(symbol.strip().upper())
-
-    args = ["src/main.py", "import"]
-    if full:
-        args.append("--full")
-    else:
-        args.extend(["--category", category])
-    return _run_cmd_streaming(args)
+# ── Runner tools (shell-command wrappers) — defined in runners.py ─────────────
+# Re-exported here so existing `from src.tools.skills_tools import X` calls
+# keep working without change.
+from src.tools.runners import (  # noqa: E402
+    run_goldbees_pipeline,
+    run_daily_signal_composite,
+    run_macro_scanner,
+    run_etf_news_sentiment,
+    run_dsp_multi_asset_importer,
+    run_data_engineering_importer,
+    run_comex_analysis,
+    run_whale_tracker,
+    run_dsp_multi_asset_comparison,
+    run_fund_mom_returns,
+    run_market_indicators,
+)
 
 
 def import_symbol_data_impl(
@@ -381,53 +208,11 @@ def import_symbol_data(
     return import_symbol_data_impl(symbol, days, start_date, end_date)
 
 
-@tool
-def run_risk_governor_analysis() -> str:
-    """
-    Compute GARCH-based position sizing and volatility targeting decision for GOLDBEES.
-    Use this when asked about GOLDBEES position sizing, GARCH volatility, risk targeting, or risk model output.
-    """
-    python_code = """
-import sys; sys.path.insert(0,'.')
-from src.tools.risk_governor import compute_position_weight, explain_decision, vol_target_for
-from src.db.pool import get_pool
-import pandas as pd, warnings
-warnings.filterwarnings('ignore')
-try:
-    price_df = get_pool().query_df('''
-        SELECT trade_date,
-               toFloat64(argMax(open,   imported_at)) AS open,
-               toFloat64(argMax(high,   imported_at)) AS high,
-               toFloat64(argMax(low,    imported_at)) AS low,
-               toFloat64(argMax(close,  imported_at)) AS close,
-               toFloat64(argMax(volume, imported_at)) AS volume
-        FROM market_data.daily_prices
-        WHERE symbol='GOLDBEES' AND category='etfs'
-        GROUP BY trade_date ORDER BY trade_date DESC LIMIT 300
-    ''')
-    price_df['trade_date'] = pd.to_datetime(price_df['trade_date'])
-    price_df = price_df.sort_values('trade_date').reset_index(drop=True)
-    from src.ml.anomaly import run_composite_anomaly
-    df_r, _, _ = run_composite_anomaly(price_df)
-    garch_vol = float(df_r['garch_vol'].dropna().iloc[-1])
-    regime    = str(df_r['regime'].iloc[-1])
-    latest    = float(df_r['close'].iloc[-1])
-    ema50     = float(price_df['close'].ewm(span=50, adjust=False).mean().iloc[-1])
-    below_ema = latest < ema50
-except Exception as e:
-    garch_vol = 16.5; regime = '✅ Normal'; below_ema = False
-    print(f'Warning: using defaults ({e})')
-
-vol_target = vol_target_for('GOLDBEES')
-d = compute_position_weight(
-    garch_annual_vol_pct=garch_vol,
-    regime=regime,
-    vol_target_pct=vol_target,
-    price_below_ema50=below_ema,
+# ── Gold/GARCH domain tools — defined in market/gold.py ──────────────────────
+from src.tools.market.gold import (  # noqa: E402
+    run_risk_governor_analysis,
+    explain_price_anomalies,
 )
-print(explain_decision(d))
-"""
-    return _run_cmd(["-c", python_code])
 
 
 @tool
@@ -522,174 +307,6 @@ def query_clickhouse_db(sql_query: str) -> str:
             f"SQL attempted:\n```sql\n{clean_query}\n```"
         )
         return f"Error executing ClickHouse query: {e}{hint}"
-
-
-def _summarize_whale_tracker_output(output: str) -> str:
-    """Parses raw whale tracker output and appends a concise summary of main accumulation/trim actions as Markdown tables."""
-    accumulations = []
-    trims = []
-    
-    current_fund = ""
-    for line in output.splitlines():
-        line_str = line.strip()
-        if "Multi Asset" in line_str:
-            current_fund = line_str.split("(")[0].strip()
-            
-        # Parse lines containing changes
-        if "%" in line_str and ("+" in line_str or "-" in line_str):
-            parts = line_str.split()
-            if len(parts) >= 5:
-                change_str = parts[-1].rstrip("%")
-                try:
-                    change_val = float(change_str)
-                    pct_indices = [i for i, p in enumerate(parts) if "%" in p]
-                    if len(pct_indices) >= 2:
-                        prev_pct_idx = pct_indices[-2]
-                        theme_idx = 1 if parts[0].startswith(("🥇", "🥈", "⚛️", "🛢️", "🏗️")) else 0
-                        theme = " ".join(parts[:theme_idx+1])
-                        security = " ".join(parts[theme_idx+1:prev_pct_idx])
-                        
-                        if change_val > 0.05:
-                            accumulations.append((change_val, security, theme, current_fund))
-                        elif change_val < -0.05:
-                            trims.append((change_val, security, theme, current_fund))
-                except ValueError:
-                    continue
-                    
-    # Sort by absolute change magnitude
-    accumulations.sort(key=lambda x: x[0], reverse=True)
-    trims.sort(key=lambda x: x[0])
-    
-    summary = "\n\n### 🐋 Whale Tracker Concise Summary\n\n"
-    summary += "#### Top Accumulations (Increasing Weight)\n\n"
-    summary += "| Fund | Theme | Security | Change |\n"
-    summary += "| :--- | :--- | :--- | ---: |\n"
-    if accumulations:
-        for change_val, security, theme, fund in accumulations[:5]:
-            summary += f"| {fund} | {theme} | {security} | {change_val:+.2f}% |\n"
-    else:
-        summary += "| None detected | | | |\n"
-        
-    summary += "\n#### Top Trims (Reducing Weight)\n\n"
-    summary += "| Fund | Theme | Security | Change |\n"
-    summary += "| :--- | :--- | :--- | ---: |\n"
-    if trims:
-        for change_val, security, theme, fund in trims[:5]:
-            summary += f"| {fund} | {theme} | {security} | {change_val:+.2f}% |\n"
-    else:
-        summary += "| None detected | | | |\n"
-
-    # Extract sector/theme trend allocations if present in the output
-    if "Unified Macro Theme Allocations" in output:
-        sub = output.split("Unified Macro Theme Allocations")[1]
-        if "High-Conviction Equity Cross-Ownership" in sub:
-            sub = sub.split("High-Conviction Equity Cross-Ownership")[0]
-            
-        themes = []
-        latest_weights = []
-        flow_changes = []
-        
-        for line in sub.splitlines():
-            if "%" in line:
-                line_clean = line
-                for emoji in ["🥈", "🥇", "⚛️", "🛢️", "🏗️"]:
-                    line_clean = line_clean.replace(emoji, "")
-                parts = line_clean.split()
-                if len(parts) >= 4:
-                    theme_name = parts[0]
-                    if theme_name in ["Silver", "Gold", "Nuclear/Grid", "Energy", "Infra"]:
-                        try:
-                            latest_w = float(parts[-2].replace("%", "").strip())
-                            flow_c = float(parts[-1].replace("%", "").strip())
-                            themes.append(theme_name)
-                            latest_weights.append(latest_w)
-                            flow_changes.append(flow_c)
-                        except ValueError:
-                            pass
-        
-        if themes:
-            try:
-                import plotext as plt
-                import re
-                ansi_escape = re.compile(r"\x1B(?:[@-Z\\-_]|\[[0-?]*[ -/]*[@-~])")
-                
-                # Plot Combined Latest Weights
-                plt.clear_figure()
-                plt.bar(themes, latest_weights, orientation="horizontal")
-                plt.title("Combined Latest Weights by Sector/Theme (%)")
-                plt.plot_size(70, 15)
-                latest_chart = plt.build()
-                latest_chart_clean = ansi_escape.sub("", latest_chart)
-                
-                # Plot Net Flow Change
-                plt.clear_figure()
-                plt.bar(themes, flow_changes, orientation="horizontal")
-                plt.title("Net Flow Change by Sector/Theme (%)")
-                plt.plot_size(70, 15)
-                flow_chart = plt.build()
-                flow_chart_clean = ansi_escape.sub("", flow_chart)
-                
-                summary += "\n\n#### Combined Latest Weights by Sector/Theme (%)\n"
-                summary += f"```text\n{latest_chart_clean}\n```\n"
-                summary += "\n\n#### Net Flow Change by Sector/Theme (%)\n"
-                summary += f"```text\n{flow_chart_clean}\n```\n"
-            except Exception as e:
-                summary += f"\n\n*(Note: Could not generate ASCII charts: {e})*\n"
-                
-    return summary
-
-
-@tool
-def run_whale_tracker() -> str:
-    """
-    Track weight shifts and institutional moves in core macro themes (Gold, Silver, Nuclear/Grid, Energy, Infra)
-    across all 7 major multi-asset funds: Nippon India, DSP Multi Asset, DSP Omni FoF, Bajaj, Quant, and ICICI.
-    Use this to identify what large institutional multi-asset funds are accumulating or trimming.
-    """
-    raw_output = _run_cmd(["src/scripts/market/whale_tracker.py"])
-    return _summarize_whale_tracker_output(raw_output)
-
-
-@tool
-def run_dsp_multi_asset_comparison() -> str:
-    """
-    Run a comparative analysis between DSP Multi Asset Allocation Fund (Standard) and DSP Multi Asset Omni FoF.
-    Compares asset class weights, structures, taxation, and Netra Quant framework strategies.
-    """
-    return _run_cmd(["src/scripts/dsp/compare_dsp_multi_asset.py"])
-
-
-@tool
-def run_fund_mom_returns(scheme_code: str | None = None, search_query: str | None = None, months: int = 12) -> str:
-    """
-    Analyze Month-over-Month (MoM) NAV returns for any Indian mutual fund.
-    You must provide either a scheme_code or a search_query.
-    Args:
-        scheme_code: The scheme code of the fund (e.g., '152056' for DSP Multi Asset).
-        search_query: Search string to look up the fund scheme code if not known (e.g., 'DSP Multi Asset').
-        months: Number of months of history to fetch (default 12).
-    """
-    if not scheme_code and not search_query:
-        return "Error: You must provide either scheme_code or search_query."
-    
-    args = ["src/scripts/portfolio/fund_mom_returns.py", "--months", str(months)]
-    if scheme_code:
-        args.extend(["--scheme", str(scheme_code)])
-    elif search_query:
-        args.extend(["--search", search_query])
-        
-    return _run_cmd(args)
-
-
-@tool
-def run_comex_analysis() -> str:
-    """
-    Run COMEX commodity pre-market signal analysis.
-    This fetches live spot prices from gold-api.com for Gold (XAU), Silver (XAG), Copper (HG),
-    compares against previous-day Yahoo Finance futures closes, and classifies signals.
-    Use this when asked for COMEX gold/silver/copper prices, COMEX signals, or commodity pre-market trends.
-    """
-    return _run_cmd(["src/main.py", "comex"])
 
 
 @tool
@@ -825,25 +442,6 @@ def run_premium_alerts(
     return _run_cmd(args)
 
 
-# Unified list of core skill tools
-SKILLS_TOOLS = [
-    run_goldbees_pipeline,
-    run_daily_signal_composite,
-    run_macro_scanner,
-    run_etf_news_sentiment,
-    run_dsp_multi_asset_importer,
-    run_data_engineering_importer,
-    import_symbol_data,
-    run_risk_governor_analysis,
-    query_clickhouse_db,
-    run_whale_tracker,
-    run_dsp_multi_asset_comparison,
-    run_fund_mom_returns,
-    run_comex_analysis,
-    run_premium_alerts,
-]
-
-
 @tool
 def run_deepdive_analysis(ticker: str, section: str | None = None, skip_fetch: bool = False) -> str:
     """
@@ -899,394 +497,28 @@ def run_deepdive_analysis(ticker: str, section: str | None = None, skip_fetch: b
     return f"Deep-dive analysis executed.\n\nCommand Log:\n{cmd_output}{preview}"
 
 
-@tool
-def run_market_indicators() -> str:
-    """
-    Run the index valuation, market breadth, sector rotation, and macro indicators scorecard.
-    Use this when the user asks for Nifty 50 or Nifty 500 P/E, P/B ratios, market breadth (Advances/Declines,
-    percentage of stocks above 50/200 DMA), sector rotation rankings, rupee stress (USDINR deviation),
-    or gold ETF (SPDR GLD) whale flows.
-    """
-    return _run_cmd(["src/scripts/portfolio/market_indicators.py"])
-
-
-@tool
-def explain_price_anomalies(symbol: str, exchange: str | None = "NSE", days: int = 60) -> str:
-    """
-    Scan price history for GOLDBEES or any asset to identify return anomalies (daily return outlier shocks > 2%)
-    in the last N days, automatically query historical news for those dates, and explain the causes.
-    Always call `plot_price_chart` in parallel with this tool to visually display the price trend.
-    Use this when the user asks to explain GOLDBEES price anomalies, chart spikes, or sudden drops.
-    """
-    import pandas as pd
-    import numpy as np
-    from datetime import datetime
-    from src.db.pool import query_df
-    from src.tools.news_search import search_financial_news
-    from src.utils.symbol_mapper import get_company_name
-
-    symbol_upper = symbol.strip().upper()
-    exchange_val = (exchange or "NSE").strip().upper()
-
-    # 1. Fetch full OHLCV from ClickHouse (GARCH needs open/high/low/close/volume + ≥60 rows)
-    df = pd.DataFrame()
-    try:
-        q = f"""
-            SELECT trade_date,
-                   toFloat64(argMax(open,   imported_at)) AS open,
-                   toFloat64(argMax(high,   imported_at)) AS high,
-                   toFloat64(argMax(low,    imported_at)) AS low,
-                   toFloat64(argMax(close,  imported_at)) AS close,
-                   toFloat64(argMax(volume, imported_at)) AS volume
-            FROM market_data.daily_prices FINAL
-            WHERE symbol = '{symbol_upper}'
-            GROUP BY trade_date ORDER BY trade_date ASC
-        """
-        df = query_df(q)
-    except Exception as e:
-        logger.warning("ClickHouse query failed, falling back to yfinance: %s", e)
-
-    if df.empty:
-        try:
-            import yfinance as yf
-            if symbol_upper.endswith("=F") or "=F" in symbol_upper:
-                ticker_sym = symbol_upper
-            else:
-                suffix = ".BO" if exchange_val == "BSE" else ".NS"
-                ticker_sym = f"{symbol_upper}{suffix}"
-            hist = yf.Ticker(ticker_sym).history(period="2y")
-            if not hist.empty:
-                df = hist.reset_index()[["Date", "Open", "High", "Low", "Close", "Volume"]]
-                df.columns = ["trade_date", "open", "high", "low", "close", "volume"]
-                df["trade_date"] = pd.to_datetime(df["trade_date"]).dt.tz_localize(None)
-        except Exception as e:
-            return f"Error: Could not retrieve price history for {symbol_upper}: {e}"
-
-    if df.empty:
-        return f"Error: Price history for {symbol_upper} is empty."
-
-    df = df.sort_values("trade_date").reset_index(drop=True)
-    df["volume"] = df["volume"].fillna(0.0)
-    # Display rolling indicators computed on the full series before any slicing
-    df["daily_return"]  = df["close"].pct_change() * 100
-    df["volatility_20d"] = df["daily_return"].rolling(20, min_periods=1).std()
-    df["volume_ma20"]    = df["volume"].rolling(20, min_periods=1).mean()
-
-    cutoff_date = pd.Timestamp(datetime.now().date()) - pd.Timedelta(days=days)
-
-    if df[df["trade_date"] >= cutoff_date].empty:
-        return f"No price data found in the last {days} days for {symbol_upper}."
-
-    # 2. Try the GARCH composite pipeline (needs ≥60 rows + arch library)
-    # Runs on FULL history so GARCH has a long series; then flags are filtered to the window.
-    composite_ok = False
-    regime_map: dict = {}   # trade_date (Timestamp) → regime label
-    finalz_map: dict = {}   # trade_date (Timestamp) → final_z float
-    threshold = 2.5         # default display threshold when composite succeeds
-
-    is_gold = "GOLD" in symbol_upper
-
-    try:
-        if len(df) < 60:
-            raise ValueError(f"Only {len(df)} rows — need ≥60 for GARCH")
-
-        # Load cross-asset features used by the composite pipeline (gold only for COT)
-        df_cot: "pd.DataFrame | None" = None
-        df_fx:  "pd.DataFrame | None" = None
-
-        if is_gold:
-            try:
-                _cot = query_df("SELECT report_date, mm_net, open_interest FROM market_data.cot_gold")
-                if not _cot.empty:
-                    _cot["report_date"] = pd.to_datetime(_cot["report_date"])
-                    df_cot = _cot
-            except Exception as _e:
-                logger.warning("COT fetch failed (non-fatal): %s", _e)
-
-        try:
-            _fx = query_df(
-                "SELECT symbol, trade_date, toFloat64(close) AS close "
-                "FROM market_data.fx_rates FINAL WHERE symbol = 'USDINR'"
-            )
-            if not _fx.empty:
-                _fx["trade_date"] = pd.to_datetime(_fx["trade_date"])
-                df_fx = _fx
-        except Exception as _e:
-            logger.warning("FX fetch failed (non-fatal): %s", _e)
-
-        from src.ml.anomaly import run_composite_anomaly
-        df_result, df_flagged, _ = run_composite_anomaly(
-            df[["trade_date", "open", "high", "low", "close", "volume"]].copy(),
-            df_cot=df_cot,
-            df_fx=df_fx,
-        )
-
-        for _, r in df_result.iterrows():
-            ts = pd.Timestamp(r["trade_date"]).normalize()
-            regime_map[ts] = r.get("regime", "—")
-            finalz_map[ts] = r.get("final_z")
-
-        # Filter flagged dates to the recent window
-        recent_flagged = df_flagged[
-            pd.to_datetime(df_flagged["trade_date"]) >= cutoff_date
-        ]["trade_date"].apply(lambda d: pd.Timestamp(d).normalize())
-        anomalies = df[df["trade_date"].apply(
-            lambda d: pd.Timestamp(d).normalize()
-        ).isin(set(recent_flagged))].copy()
-        composite_ok = True
-
-    except Exception as _e:
-        logger.warning("Composite anomaly failed, using naive threshold fallback: %s", _e)
-        df_recent = df[df["trade_date"] >= cutoff_date].copy()
-        std_ret = df_recent["daily_return"].std()
-        threshold = max(2.0, 2.5 * std_ret) if not pd.isna(std_ret) else 2.0
-        anomalies = df_recent[df_recent["daily_return"].abs() >= threshold].copy()
-
-    if anomalies.empty:
-        detection_note = (
-            "GARCH composite (Final Z > 2.5)" if composite_ok
-            else f"naive threshold ({threshold:.2f}%)"
-        )
-        return (
-            f"No price anomalies detected for {symbol_upper} in the last {days} days "
-            f"({detection_note})."
-        )
-
-    anomalies = anomalies.sort_values("trade_date", ascending=False)
-
-    # 3. Pre-load repo readers for forward ML/signal context (Part B)
-    repo = None
-    try:
-        from src.db.pool import get_pool
-        from src.db.repository import MarketDataRepository
-        repo = MarketDataRepository(get_pool())
-    except Exception as _e:
-        logger.warning("Could not initialise repository for forward context: %s", _e)
-
-    # 4. Check if stock; fetch quarterly results for earnings correlation
-    is_stock = False
-    quarterly_results = None
-    try:
-        from src.tools.inav_fetcher import is_etf
-        is_stock = not is_etf(symbol_upper)
-    except Exception as e:
-        logger.warning("Could not determine if %s is an ETF: %s", symbol_upper, e)
-
-    if is_stock:
-        try:
-            from src.tools.earnings_scraper import get_quarterly_results
-            res = get_quarterly_results.invoke({"input_str": f"{symbol_upper}:{exchange_val}"})
-            if res and "error" not in res:
-                quarterly_results = res
-        except Exception as e:
-            logger.warning("Failed to fetch quarterly results for %s: %s", symbol_upper, e)
-
-    def format_volume(vol: float) -> str:
-        if vol >= 10000000:
-            return f"{vol / 10000000:.2f} Cr"
-        elif vol >= 100000:
-            return f"{vol / 100000:.2f} L"
-        elif vol >= 1000:
-            return f"{vol / 1000:.1f}k"
-        else:
-            return f"{vol:,.0f}"
-
-    detection_method = "GARCH composite (Final Z > 2.5)" if composite_ok else f"naive threshold (≥{threshold:.2f}%)"
-    output = []
-    output.append(f"### 🔍 Price Anomaly & News Correlation Report: {symbol_upper}")
-    output.append(f"Detected **{len(anomalies)}** anomaly dates in the last {days} days ({detection_method}):\n")
-
-    # Summary table — includes Regime + Final Z when composite succeeded
-    output.append("| Date | Close Price (₹) | Daily Return (%) | 20d Volatility | Volume (Spike) | Regime | Final Z | News Search Query |")
-    output.append("| :--- | :--- | :--- | :--- | :--- | :--- | :--- | :--- |")
-
-    anomaly_queries = []
-    for _, row in anomalies.iterrows():
-        t_date     = row["trade_date"]
-        t_date_key = pd.Timestamp(t_date).normalize()
-        date_str   = t_date_key.strftime("%Y-%m-%d")
-        daily_ret  = row["daily_return"]
-        close_px   = row["close"]
-        vol_20d    = row["volatility_20d"]
-        volume     = row["volume"]
-        vol_ma     = row["volume_ma20"]
-        vol_ratio  = volume / vol_ma if vol_ma > 0 else 1.0
-
-        vol_str      = format_volume(volume)
-        vol_spike_str = f"{vol_str} ({vol_ratio:.1f}x)"
-        vol_20d_str  = f"{vol_20d:.2f}%" if not pd.isna(vol_20d) else "N/A"
-
-        regime  = regime_map.get(t_date_key, "—")
-        fz_val  = finalz_map.get(t_date_key)
-        fz_str  = f"{fz_val:+.2f}" if fz_val is not None else "N/A"
-
-        company_name = get_company_name(symbol_upper)
-        if "GOLD" in symbol_upper:
-            query = "gold price India custom duty import tax"
-        elif "SILVER" in symbol_upper:
-            query = "silver price India customs tax"
-        elif symbol_upper in ["NIFTYBEES", "BANKBEES", "JUNIORBEES", "MID150BEES"]:
-            query = "Nifty stock market India news"
-        else:
-            query = f"{company_name or symbol_upper} stock news"
-
-        output.append(
-            f"| {date_str} | {close_px:.2f} | {daily_ret:+.2f}% | {vol_20d_str} | {vol_spike_str} | {regime} | {fz_str} | `{query}` |"
-        )
-        anomaly_queries.append((date_str, close_px, daily_ret, vol_20d_str, vol_spike_str, regime, fz_str, query))
-
-    output.append("\n" + "─" * 40 + "\n")
-    output.append("### 📰 Detailed Date-by-Date News Correlation:\n")
-
-    for date_str, close_px, daily_ret, vol_20d_str, vol_spike_str, regime, fz_str, query in anomaly_queries:
-        output.append(f"#### 📅 Date: **{date_str}** | Close: ₹{close_px:.2f} | Daily Return: **{daily_ret:+.2f}%**")
-        if regime != "—":
-            output.append(f"- **GARCH Regime:** {regime}")
-        if fz_str != "N/A":
-            output.append(f"- **Composite Final Z:** {fz_str}")
-        output.append(f"- **20-Day Rolling Daily Volatility:** {vol_20d_str}")
-        output.append(f"- **Trading Volume:** {vol_spike_str}")
-        output.append(f"*Searching news for: \"{query}\" on {date_str}...*")
-
-        try:
-            news_output = search_financial_news.invoke({"query": query, "max_results": 3, "target_date": date_str})
-            if "No news found" in news_output:
-                fallback_query = f"{symbol_upper} share price news"
-                news_output = search_financial_news.invoke({"query": fallback_query, "max_results": 3, "target_date": date_str})
-            output.append(news_output)
-
-            # Flag neutral-news + large-move divergence (policy surprise / pre-positioning)
-            if abs(daily_ret) >= 3.0 and "neutral" in news_output.lower():
-                output.append(
-                    "\n> ⚠️ **Divergence signal:** Neutral news sentiment on a high-magnitude move — "
-                    "possible policy surprise or pre-positioning before public announcement."
-                )
-
-            # Earnings correlation for stocks
-            if quarterly_results:
-                news_lower = news_output.lower()
-                earnings_keywords = ["result", "earning", "profit", "revenue", "q1", "q2", "q3", "q4", "sales"]
-                if any(kw in news_lower for kw in earnings_keywords):
-                    output.append("\n**📊 Correlated Quarterly Financial Results:**")
-                    output.append(f"- **Reporting Period:** {quarterly_results.get('period', 'N/A')}")
-
-                    rev_yoy = quarterly_results.get('revenue_yoy_pct')
-                    rev_cr  = quarterly_results.get('revenue_cr')
-                    rev_str = f"₹{rev_cr:.2f} Cr" if isinstance(rev_cr, (int, float)) else str(rev_cr)
-                    if isinstance(rev_yoy, (int, float)):
-                        output.append(f"- **Revenue:** {rev_str} ({rev_yoy:+.2f}% YoY)")
-                    else:
-                        output.append(f"- **Revenue:** {rev_str} ({rev_yoy} YoY)")
-
-                    prof_yoy = quarterly_results.get('profit_yoy_pct')
-                    prof_cr  = quarterly_results.get('net_profit_cr')
-                    prof_str = f"₹{prof_cr:.2f} Cr" if isinstance(prof_cr, (int, float)) else str(prof_cr)
-                    if isinstance(prof_yoy, (int, float)):
-                        output.append(f"- **Net Profit:** {prof_str} ({prof_yoy:+.2f}% YoY)")
-                    else:
-                        output.append(f"- **Net Profit:** {prof_str} ({prof_yoy} YoY)")
-
-                    eps     = quarterly_results.get('eps')
-                    eps_yoy = quarterly_results.get('eps_yoy_pct')
-                    eps_str = f"₹{eps:.2f}" if isinstance(eps, (int, float)) else str(eps)
-                    if isinstance(eps_yoy, (int, float)):
-                        output.append(f"- **EPS:** {eps_str} ({eps_yoy:+.2f}% YoY)")
-                    else:
-                        output.append(f"- **EPS:** {eps_str} ({eps_yoy} YoY)")
-
-                    if quarterly_results.get('guidance'):
-                        output.append(f"- **Guidance:** {quarterly_results.get('guidance')}")
-                    output.append(f"- **Source:** [Screener/Yahoo]({quarterly_results.get('source_url', '#')})")
-
-        except Exception as exc:
-            output.append(f"  ❌ News search failed: {exc}")
-
-        # Part B — forward ML/signal context as-of each anomaly date
-        if repo is not None:
-            try:
-                ml_pred = repo.ml_prediction_asof(date_str)
-                sig     = repo.signal_composite_asof(symbol_upper, date_str)
-                if ml_pred or sig:
-                    output.append("\n**📡 What the models said on this date:**")
-                if ml_pred:
-                    signal_label = ml_pred.get("regime_signal", "N/A")
-                    prob         = ml_pred.get("prob_up", 0.0)
-                    exp_ret      = ml_pred.get("expected_return_pct", 0.0)
-                    # Characterise expected direction for narration
-                    direction = "mean-reversion (bearish 5d expectation)" if exp_ret < 0 else "continuation (bullish 5d expectation)"
-                    output.append(
-                        f"- **ML (5d forecast):** `{signal_label}` | prob_up={prob:.0%} | "
-                        f"expected_return={exp_ret:+.2f}% → {direction}"
-                    )
-                if sig:
-                    action     = sig.get("action", "N/A")
-                    score      = sig.get("composite_score", 0.0)
-                    anom_flag  = sig.get("anomaly_flag", "")
-                    confirmed  = action in ("BUY", "WATCH_LONG")
-                    verdict    = "✅ Signal confirmed shock" if confirmed else "⚠️ Signal contradicted by shock"
-                    output.append(
-                        f"- **Composite signal:** `{action}` (score={score:.1f}, anomaly_flag={anom_flag}) — {verdict}"
-                    )
-            except Exception as _e:
-                logger.warning("Forward context lookup failed for %s on %s: %s", symbol_upper, date_str, _e)
-
-        output.append("\n" + "─" * 40 + "\n")
-
-    # Append COMEX price chart if it's a Gold or Silver ETF
-    comex_symbol = None
-    if "GOLD" in symbol_upper:
-        comex_symbol = "GC=F"
-    elif "SILVER" in symbol_upper:
-        comex_symbol = "SI=F"
-
-    if comex_symbol:
-        try:
-            from src.tools.chart_tools import plot_price_chart
-            comex_chart = plot_price_chart.invoke({"symbol": comex_symbol, "days": days})
-            output.append("### 📈 Correlated COMEX Futures Price Chart:")
-            output.append(f"To assist in visual correlation, here is the price chart for the underlying COMEX futures contract (**{comex_symbol}**):\n")
-            output.append("```text")
-            output.append(comex_chart)
-            output.append("```")
-            output.append("\n" + "─" * 40 + "\n")
-        except Exception as e:
-            logger.warning("Failed to append COMEX chart for %s: %s", symbol_upper, e)
-
-    # Append GARCH Volatility Chart if GARCH data exists in weight_checkpoints
-    try:
-        from src.tools.chart_tools import plot_garch_volatility_chart
-        garch_chart = plot_garch_volatility_chart.invoke({"symbol": symbol_upper, "days": days})
-        if "No GARCH data found" not in garch_chart:
-            output.append("### 📊 GARCH Annualised Volatility Chart:")
-            output.append(f"Here is the GARCH(1,1) annualised volatility trend for **{symbol_upper}** over the last {days} days:\n")
-            output.append("```text")
-            output.append(garch_chart)
-            output.append("```")
-            output.append("\n" + "─" * 40 + "\n")
-    except Exception as e:
-        logger.warning("Failed to append GARCH volatility chart for %s: %s", symbol_upper, e)
-
-    return "\n".join(output)
-
-
-# Unified list of core skill tools
+# ── Canonical tool list ────────────────────────────────────────────────────────
+# Single source of truth — all other lists are subsets of this.
 SKILLS_TOOLS = [
+    # Runners (shell wrappers) — defined in runners.py, re-exported above
     run_goldbees_pipeline,
     run_daily_signal_composite,
     run_macro_scanner,
     run_etf_news_sentiment,
     run_dsp_multi_asset_importer,
     run_data_engineering_importer,
-    import_symbol_data,
-    get_live_inav,
-    run_risk_governor_analysis,
-    query_clickhouse_db,
+    run_comex_analysis,
     run_whale_tracker,
     run_dsp_multi_asset_comparison,
     run_fund_mom_returns,
-    run_comex_analysis,
+    run_market_indicators,
+    # Gold/GARCH domain — defined in market/gold.py, re-exported above
+    run_risk_governor_analysis,
+    explain_price_anomalies,
+    # General-purpose tools defined in this file
+    import_symbol_data,
+    get_live_inav,
+    query_clickhouse_db,
     run_premium_alerts,
     run_deepdive_analysis,
-    run_market_indicators,
-    explain_price_anomalies,
 ]
